Remove file case normalization
The FileResolver methods were normalizing the case
of files and doing expensive operations like getting
the canonical file for this. The exact background of
this is no longer clear. It has been around since Gradle
1.0, but doesn't serve any correctness purpose. For correctness
one would have to compare the canonical file, so just
normalizing the case isn't enough.

Our best guess is that this was introduced to increase the
potential for cache hits on case-insensitve file systems in
the unusual case that the user provides a String that doesn't
match the actual case of the file on disk. Another theory is
that it makes error messages easier to understand, as they
match the file's name on disk exactly.

The method has been changed to only remove segments like /../,
which is cheap and can be done without using the file system.
This saves a lot of IO during configuration time and when resolving
file dependencies.