Remove file existence check on UriTextResource
If the resource was given in the form of a file,
but that file does not exist, it is invalid in any
case. Hiding that fact by returning "null" does not
help, other methods like getReader will still fail.

Checking for existence slows down the common case of
a script that does exist.